Add message count to level

+++ src/main/java/org/geysermc/discordbot/util/BotHelpers.java
@@ -232,8 +232,8 @@ public class BotHelpers {
      * @param input Time string to parse
      * @return Time in seconds
      */
-    public static long parseTimeString(String input) {
-        long result = 0;
+    public static int parseTimeString(String input) {
+        int result = 0;
         String number = "";
         for (int i = 0; i < input.length(); i++) {
             char c = input.charAt(i);
@@ -247,11 +247,11 @@ public class BotHelpers {
         return result;
     }
 
-    private static long convert(int value, char unit) {
-        switch(unit) {
-            case 'd' : return (long) value * 60 * 60 * 24;
-            case 'h' : return (long) value * 60 * 60;
-            case 'm' : return (long) value * 60;
+    private static int convert(int value, char unit) {
+        switch (unit) {
+            case 'd' : return value * 60 * 60 * 24;
+            case 'h' : return value * 60 * 60;
+            case 'm' : return value * 60;
             case 's' : return value;
         }
         return 0;
